+++ tests/bgen/bgen-tests.csproj
@@ -46,6 +46,9 @@
     <Compile Include="..\generator\NullabilityContextTests.cs">
       <Link>NullabilityContextTests.cs</Link>
     </Compile>
+    <Compile Include="..\generator\StringExtensionTests.cs">
+      <Link>StringExtensionTests.cs</Link>
+    </Compile>
     <Compile Include="..\mtouch\Cache.cs">
       <Link>Cache.cs</Link>
     </Compile>
