+++ tests/generator/generator-tests.csproj
@@ -64,6 +64,7 @@
       <Link>BinLog.cs</Link>
     </Compile>
     <Compile Include="NullabilityContextTests.cs" />
+    <Compile Include="StringExtensionTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="tests\is-direct-binding.cs" />
